Extracted environment context provider builders from conftest.py and removed unused cli fixtures
