+++ test/integration/conftest.py
@@ -1,107 +1,49 @@
-import contextlib
 import io
-from test.integration.helpers import normalize_request_name
-from typing import (
-    Any,
-    Callable,
-    ContextManager,
-    Dict,
-    Generator,
-    Iterator,
-    List,
-    Optional,
-)
+from typing import Iterator
 
 import pytest
-from typing_extensions import TypeAlias  # Needed for Python3.9
 
 from exasol_integration_test_docker_environment.test.get_test_container_content import (
     get_test_container_content,
 )
-from exasol_integration_test_docker_environment.testing import (
-    luigi_utils,
-    utils,
-)
 from exasol_integration_test_docker_environment.testing.api_test_environment import (
     ApiTestEnvironment,
+)
+from exasol_integration_test_docker_environment.testing.api_test_environment_context_provider import (
+    ApiContextProvider,
+    build_api_context_provider,
+    build_api_context_provider_with_test_container,
+    build_api_isolation,
+)
+from exasol_integration_test_docker_environment.testing.cli_test_environment_context_provider import (
+    CliContextProvider,
+    build_cli_context_provider,
+    build_cli_isolation,
 )
 from exasol_integration_test_docker_environment.testing.exaslct_docker_test_environment import (
     ExaslctDockerTestEnvironment,
 )
 from exasol_integration_test_docker_environment.testing.exaslct_test_environment import (
     ExaslctTestEnvironment,
-    SpawnedTestEnvironments,
 )
-
-
-@contextlib.contextmanager
-def _build_cli_isolation(request) -> Iterator[ExaslctTestEnvironment]:
-    testname = normalize_request_name(request.node.name)
-    environment = ExaslctTestEnvironment(
-        test_object=None,
-        executable="itde",
-        clean_images_at_close=True,
-        name=testname,
-    )
-    yield environment
-    utils.close_environments(environment)
 
 
 @pytest.fixture
 def cli_isolation(request) -> Iterator[ExaslctTestEnvironment]:
-    with _build_cli_isolation(request) as environment:
+    with build_cli_isolation(request) as environment:
         yield environment
-
-
-@pytest.fixture(scope="module")
-def cli_isolation_module(request) -> Iterator[ExaslctTestEnvironment]:
-    with _build_cli_isolation(request) as environment:
-        yield environment
-
-
-@contextlib.contextmanager
-def _build_api_isolation(request) -> Iterator[ApiTestEnvironment]:
-    testname = normalize_request_name(request.node.name)
-    environment = ApiTestEnvironment(test_object=None, name=testname)
-    yield environment
-    utils.close_environments(environment)
-    luigi_utils.clean(environment.docker_repository_name)
 
 
 @pytest.fixture
 def api_isolation(request) -> Iterator[ApiTestEnvironment]:
-    with _build_api_isolation(request) as environment:
+    with build_api_isolation(request) as environment:
         yield environment
 
 
 @pytest.fixture(scope="module")
 def api_isolation_module(request) -> Iterator[ApiTestEnvironment]:
-    with _build_api_isolation(request) as environment:
+    with build_api_isolation(request) as environment:
         yield environment
-
-
-CliContextProvider: TypeAlias = Callable[
-    [Optional[str], Optional[List[str]]], ContextManager[SpawnedTestEnvironments]
-]
-
-
-def _build_cli_context_provider(
-    test_environment: ExaslctTestEnvironment,
-) -> CliContextProvider:
-    @contextlib.contextmanager
-    def create_context(
-        name: Optional[str] = None,
-        additional_parameters: Optional[List[str]] = None,
-    ) -> Iterator[SpawnedTestEnvironments]:
-        name = name if name else test_environment.name
-        spawned = test_environment.spawn_docker_test_environments(
-            name=name,
-            additional_parameter=additional_parameters,
-        )
-        yield spawned
-        utils.close_environments(spawned)
-
-    return create_context
 
 
 @pytest.fixture
@@ -122,54 +64,7 @@ def cli_database(
         with cli_database(additional_parameters = ["--option"]):
             ...
     """
-    return _build_cli_context_provider(cli_isolation)
-
-
-@pytest.fixture(scope="module")
-def cli_database_module(
-    cli_isolation_module,
-) -> CliContextProvider:
-    """
-    Returns a method that test case implementations can use to create a
-    context with a database.
-
-    This fixture should be used on module level, in cases where one
-    database is required for all tests in the module.
-
-    The test case optionally can pass a name and additional parameters for
-    spawning the database:
-
-    def test_case(cli_database_module):
-        with cli_database_module(additional_parameters = ["--option"]):
-            ...
-    """
-    return _build_cli_context_provider(cli_isolation_module)
-
-
-ApiContextProvider: TypeAlias = Callable[
-    [Optional[str], Optional[Dict[str, Any]]],
-    ContextManager[ExaslctDockerTestEnvironment],
-]
-
-
-def _build_api_context_provider(
-    test_environment: ApiTestEnvironment,
-) -> ApiContextProvider:
-
-    @contextlib.contextmanager
-    def create_context(
-        name: Optional[str] = None,
-        additional_parameters: Optional[Dict[str, Any]] = None,
-    ) -> Generator[ExaslctDockerTestEnvironment, None, None]:
-        name = name if name else test_environment.name
-        spawned = test_environment.spawn_docker_test_environment(
-            name=name,
-            additional_parameter=additional_parameters,
-        )
-        yield spawned
-        utils.close_environments(spawned)
-
-    return create_context  # type: ignore
+    return build_cli_context_provider(cli_isolation)
 
 
 @pytest.fixture
@@ -188,7 +83,7 @@ def api_database(api_isolation: ApiTestEnvironment) -> ApiContextProvider:
         with api_database(additional_parameters = ["--option"]):
             ...
     """
-    return _build_api_context_provider(api_isolation)
+    return build_api_context_provider(api_isolation)
 
 
 @pytest.fixture(scope="module")
@@ -198,30 +93,9 @@ def api_default_database_module(
     """
     Provides a default database environment.
     """
-    provider = _build_api_context_provider(api_isolation_module)
+    provider = build_api_context_provider(api_isolation_module)
     with provider(None, None) as db:
         yield db
-
-
-def _build_api_context_provider_with_test_container(
-    test_environment: ApiTestEnvironment,
-) -> ApiContextProvider:
-    @contextlib.contextmanager
-    def create_context(
-        name: Optional[str] = None,
-        additional_parameters: Optional[Dict[str, Any]] = None,
-        test_container_content=get_test_container_content(),
-    ) -> Generator[ExaslctDockerTestEnvironment, None, None]:
-        name = name if name else test_environment.name
-        spawned = test_environment.spawn_docker_test_environment_with_test_container(
-            name=name,
-            additional_parameter=additional_parameters,
-            test_container_content=test_container_content,
-        )
-        yield spawned
-        utils.close_environments(spawned)
-
-    return create_context  # type: ignore
 
 
 @pytest.fixture
@@ -243,7 +117,10 @@ def api_database_with_test_container(
                                               test_container_content = get_test_container_content(runtime_mapping-...)):
             ...
     """
-    return _build_api_context_provider_with_test_container(api_isolation)
+    return build_api_context_provider_with_test_container(
+        api_isolation,
+        get_test_container_content(),
+    )
 
 
 @pytest.fixture(scope="module")
@@ -254,7 +131,9 @@ def api_default_database_with_test_conainer_module(
     Provides a default database + test container environment.
     """
 
-    provider = _build_api_context_provider_with_test_container(api_isolation_module)
+    provider = build_api_context_provider_with_test_container(
+        api_isolation_module, get_test_container_content()
+    )
     with provider(None, None) as db:
         yield db
